Restructure to metrics + callbacks
