+++ keras_cv/callbacks/__init__.py
@@ -1,2 +1,1 @@
-from keras_cv.metrics.coco.pycoco_utils import compute_pycoco_metrics
-from keras_cv.metrics.coco.recall import COCORecall
+from keras_cv.callbacks.pycoco_callback import PyCOCOCallback
